Give up on compiling to webassembly until memory64 is stabilized
The main blocker preventing us from compiling this library for wasm is that
there are certain problems (specifically, 2015 day 11) that rely on `usize`
being 64 bits in order to correctly implement certain traits.

It's probably possible to work around these issues, but to be honest I'd rather
just wait for memory64 to reach stage 4/5 and for wasm64-unknown-unknown to be
stabilized as at least a tier 2 target for Rust.